Collapse Char/LPStr tests into same project
