+++ src/tests/Interop/PInvoke/BestFitMapping/Assembly_Default/PInvoke_Default.cs
@@ -1,0 +1,24 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Char
+{
+    public partial class PInvoke_Default
+    {
+        public static void RunTest()
+        {
+            RunTest(bestFitMapping: true, throwOnUnmappableChar: false);
+        }
+    }
+}
+
+namespace LPStr
+{
+    public partial class PInvoke_Default
+    {
+        public static void RunTest()
+        {
+            RunTest(bestFitMapping: true, throwOnUnmappableChar: false);
+        }
+    }
+}
