+++ src/tests/Interop/PInvoke/BestFitMapping/Assembly_False_False/AssemblyInfo.cs
