+++ src/tests/Interop/PInvoke/BestFitMapping/Assembly_True_False/AssemblyInfo.cs
