+++ src/tests/Interop/PInvoke/BestFitMapping/Assembly_True_True/AssemblyInfo.cs
