--- src/tests/Interop/PInvoke/BestFitMapping/Char/Directory.Build.props
@@ -1,18 +1,0 @@
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Build.props, $(MSBuildThisFileDirectory)..))" />
-
-  <PropertyGroup>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)\PInvoke_Default.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\PInvoke_False_False.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\PInvoke_False_True.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\PInvoke_True_False.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\PInvoke_True_True.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\..\Test.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\..\Program.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)\..\TestData.cs" />
-  </ItemGroup>
-</Project>
