+++ src/tests/Interop/PInvoke/BestFitMapping/Char/PInvoke_False_False.cs
@@ -5,6 +5,8 @@ using System;
 using System.Text;
 using System.Runtime.InteropServices;
 
+namespace Char;
+
 public class PInvoke_False_False
 {
     [DllImport("Char_BestFitMappingNative", BestFitMapping = false, ThrowOnUnmappableChar = false)]
