+++ src/tests/Interop/PInvoke/BestFitMapping/Char/PInvoke_False_True.cs
@@ -5,6 +5,8 @@ using System;
 using System.Text;
 using System.Runtime.InteropServices;
 
+namespace Char;
+
 public class PInvoke_False_True
 {
     [DllImport("Char_BestFitMappingNative", BestFitMapping = false, ThrowOnUnmappableChar = true)]
