--- src/tests/Interop/PInvoke/BestFitMapping/LPStr/Assembly_False_True/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Runtime.InteropServices;
-
-[assembly: BestFitMapping(false, ThrowOnUnmappableChar = true)]
