--- src/tests/Interop/PInvoke/BestFitMapping/LPStr/Assembly_False_True/PInvoke_Default.cs
@@ -1,10 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-public partial class PInvoke_Default
-{
-    public static void RunTest()
-    {
-        RunTest(bestFitMapping: false, throwOnUnmappableChar: true);
-    }
-}
