+++ src/tests/Interop/PInvoke/BestFitMapping/LPStr/PInvoke_Default.cs
@@ -7,6 +7,8 @@ using System.Runtime.InteropServices;
 
 using static TestData;
 
+namespace LPStr;
+
 public partial class PInvoke_Default
 {
     [StructLayout(LayoutKind.Sequential)]
