+++ src/tests/Interop/PInvoke/BestFitMapping/LPStr/PInvoke_True_True.cs
@@ -7,6 +7,8 @@ using System.Runtime.InteropServices;
 
 using static TestData;
 
+namespace LPStr;
+
 public class PInvoke_True_True
 {
     [StructLayout(LayoutKind.Sequential)]
